pyproject.toml
[tool.poetry]
name = "kaizen-cloudcode"
version = "0.3.19"
description = "An intelligent coding companion that accelerates your development workflow by providing efficient assistance, enabling you to craft high-quality code more rapidly."
authors = ["Saurav Panda <saurav.panda@cloudcode.ai>"]
license = "Apache2.0"
readme = "README.md"
packages = [
    { include = "kaizen" }
]

[tool.poetry.dependencies]
python = "^3.9.0"
requests = "^2.31.0"
pyjwt = "^2.8.0"
uvicorn = "^0.29.0"
fastapi = "^0.110.0"
litellm = "^1.40.29"
flake8 = "^7.0.0"
black = "^24.3.0"
cryptography = "^42.0.5"
bs4 = "^0.0.2"
nest-asyncio = "^1.6.0"
pytest-playwright = "^0.4.4"
pip = "^24.0"
fuzzywuzzy = "^0.18.0"
llama-index-core = "^0.10.47"
llama-index-llms-openai = "^0.1.22"
llama-index-embeddings-openai = "^0.1.10"
llama-index-readers-file = "^0.1.25"
llama-index-vector-stores-postgres = "^0.1.11"
sqlalchemy = "^2.0.31"
esprima = "^4.0.1"
escodegen = "^1.0.11"
redis = "^5.0.7"
tqdm = "^4.66.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
